fix: rename filter.ts
